+++ studio/src/editors/merch-card-editor.js
@@ -1,21 +1,16 @@
 import { html, nothing } from 'lit';
-import { FragmentEditorBase } from '../fragment-editor-base.js';
 import { unsafeHTML } from 'lit/directives/unsafe-html.js';
 import '../fields/multifield.js';
 import '../fields/mnemonic-field.js';
-import { FragmentStore } from '../reactivity/fragment-store.js';
 import '../aem/aem-tag-picker-field.js';
 import './variant-picker.js';
-import '../editor-panel.js';
+import EditorPanel from '../editor-panel.js';
 
 const MODEL_PATH = '/conf/mas/settings/dam/cfm/models/card';
 
-class MerchCardEditor extends FragmentEditorBase {
+class MerchCardEditor extends EditorPanel {
     static properties = {
-        refreshing: { state: true, attribute: true },
-        disabled: { state: true, attribute: true },
-        hasChanges: { state: true, attribute: true },
-        loading: { state: true, attribute: true },
+        fragment: { type: Object, attribute: false },
     };
 
     createRenderRoot() {
@@ -24,13 +19,6 @@ class MerchCardEditor extends FragmentEditorBase {
 
     constructor() {
         super();
-        this.disabled = false;
-        this.refreshing = false;
-        this.hasChanges = false;
-        this.loading = false;
-        this.close = this.close.bind(this);
-        this.discardChanges = this.discardChanges.bind(this);
-        this.refresh = this.refresh.bind(this);
     }
 
     connectedCallback() {
@@ -63,36 +51,21 @@ class MerchCardEditor extends FragmentEditorBase {
     }
 
     render() {
-        if (this.loading)
-            return html`<div id="editor">
-                <sp-progress-circle indeterminate size="l"></sp-progress-circle>
-            </div>`;
-
-        if (this.refreshing || !this.fragment) return nothing;
-
         if (this.fragment.model.path !== MODEL_PATH) return nothing;
 
         const form = Object.fromEntries([
             ...this.fragment.fields.map((f) => [f.name, f]),
         ]);
 
-        return html`<div id="editor">
-            <editor-panel
-                .fragment=${this.fragment}
-                .source=${this.repository}
-                .bucket=${this.bucket}
-                .showToast=${this.showToast}
-                ?disabled=${this.disabled}
-                ?hasChanges=${this.hasChanges}
-            ></editor-panel>
+        return html`
             <sp-field-label for="card-variant">Variant</sp-field-label>
             <variant-picker
                 id="card-variant"
                 ?show-all="false"
                 data-field="variant"
                 default-value="${form.variant.values[0]}"
-                @input="${this.#updateFragment}"
-                @change="${this.#updateFragment}"
+                @input="${this.updateFragment}"
+                @change="${this.updateFragment}"
                 ?disabled=${this.disabled}
             ></variant-picker>
             <sp-field-label for="card-title">Title</sp-field-label>
@@ -101,7 +74,7 @@ class MerchCardEditor extends FragmentEditorBase {
                 id="card-title"
                 data-field="cardTitle"
                 value="${form.cardTitle.values[0]}"
-                @input="${this.#updateFragment}"
+                @input="${this.updateFragment}"
                 ?disabled=${this.disabled}
             ></sp-textfield>
             <sp-field-label for="card-subtitle">Subtitle</sp-field-label>
@@ -110,7 +83,7 @@ class MerchCardEditor extends FragmentEditorBase {
                 id="card-subtitle"
                 data-field="subtitle"
                 value="${form.subtitle.values[0]}"
-                @input="${this.#updateFragment}"
+                @input="${this.updateFragment}"
                 ?disabled=${this.disabled}
             ></sp-textfield>
             <sp-field-label for="card-size">Size</sp-field-label>
@@ -119,7 +92,7 @@ class MerchCardEditor extends FragmentEditorBase {
                 id="card-size"
                 data-field="size"
                 value="${form.size.values[0]}"
-                @input="${this.#updateFragment}"
+                @input="${this.updateFragment}"
                 ?disabled=${this.disabled}
             ></sp-textfield>
             <sp-field-label for="card-icon">Badge</sp-field-label>
@@ -128,7 +101,7 @@ class MerchCardEditor extends FragmentEditorBase {
                 id="card-badge"
                 data-field="badge"
                 value="${form.badge.values[0]}"
-                @input="${this.#updateFragment}"
+                @input="${this.updateFragment}"
                 ?disabled=${this.disabled}
             ></sp-textfield>
             <sp-field-label for="mnemonic">Mnemonics</sp-field-label>
@@ -144,22 +117,22 @@ class MerchCardEditor extends FragmentEditorBase {
             </mas-multifield>
             <sp-field-label for="card-icon">Background Image</sp-field-label>
             <sp-textfield
-                placeholder="Enter backgroung image URL"
+                placeholder="Enter background image URL"
                 id="background-title"
                 data-field="backgroundImage"
                 value="${form.backgroundImage.values[0]}"
-                @input="${this.#updateFragment}"
+                @input="${this.updateFragment}"
                 ?disabled=${this.disabled}
             ></sp-textfield>
             <sp-field-label for="card-icon"
                 >Background Image Alt Text</sp-field-label
             >
             <sp-textfield
-                placeholder="Enter backgroung image Alt Text"
+                placeholder="Enter background image Alt Text"
                 id="background-alt-text"
                 data-field="backgroundImageAltText"
                 value="${form.backgroundImageAltText.values[0]}"
-                @input="${this.#updateFragment}"
+                @input="${this.updateFragment}"
                 ?disabled=${this.disabled}
             ></sp-textfield>
             <sp-field-label for="horizontal"> Prices </sp-field-label>
@@ -168,7 +141,7 @@ class MerchCardEditor extends FragmentEditorBase {
                     inline
                     data-field="prices"
                     default-link-style="primary-outline"
-                    @change="${this.#updateFragment}"
+                    @change="${this.updateFragment}"
                     ?readonly=${this.disabled}
                     >${unsafeHTML(form.prices.values[0])}</rte-field
                 >
@@ -179,7 +152,7 @@ class MerchCardEditor extends FragmentEditorBase {
                     link
                     data-field="description"
                     default-link-style="secondary-link"
-                    @change="${this.#updateFragment}"
+                    @change="${this.updateFragment}"
                     ?readonly=${this.disabled}
                     >${unsafeHTML(form.description.values[0])}</rte-field
                 >
@@ -191,7 +164,7 @@ class MerchCardEditor extends FragmentEditorBase {
                     inline
                     data-field="ctas"
                     default-link-style="primary-outline"
-                    @change="${this.#updateFragment}"
+                    @change="${this.updateFragment}"
                     ?readonly=${this.disabled}
                     >${unsafeHTML(form.ctas.values[0])}</rte-field
                 >
@@ -203,21 +176,13 @@ class MerchCardEditor extends FragmentEditorBase {
                 value="${this.fragment.tags.map((tag) => tag.id).join(',')}"
                 @change=${this.#handeTagsChange}
             ></aem-tag-picker-field>
-        </div> `;
+        `;
     }
 
     #handeTagsChange(e) {
         const value = e.target.getAttribute('value');
         const newTags = value ? value.split(',') : []; // do not overwrite the tags array
         this.fragmentStore.updateField('tags', newTags);
-        this.hasChanges = true;
-    }
-
-    #updateFragment(event) {
-        const fieldName = event.target.dataset.field;
-        let value = event.target.value || event.detail?.value;
-        value = event.target.multiline ? value?.split(',') : [value ?? ''];
-        this.fragmentStore.updateField(fieldName, value);
         this.hasChanges = true;
     }
 
@@ -240,19 +205,3 @@ class MerchCardEditor extends FragmentEditorBase {
 }
 
 customElements.define('merch-card-editor', MerchCardEditor);
-
-/**
- * @returns {MerchCardEditor}
- */
-export function getMerchCardEditor() {
-    return document.querySelector('merch-card-editor');
-}
-
-/**
- * @param {FragmentStore} store
- * @param {number | undefined} x - The clientX value of the mouse event (used for positioning - optional)
- */
-export async function editFragment(store, x) {
-    const editor = getMerchCardEditor();
-    editor.editFragment(store, x);
-}
